# docs/api.yaml
...
openapi: 3.0.0
info:
  title: AI Chat Notes API
  version: 1.0.0
  description: AI 聊天聚合 + 知识库管理 API
servers:
  - url: /api
    description: API 路径前缀

# ============================================
# 通用定义
# ============================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误详情
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          description: 成功信息
      required:
        - message
...
    Model:
      type: object
      deprecated: true
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
          description: 显示名称
        api_base:
          type: string
          description: API 地址
        api_key:
          type: string
          description: API Key（仅创建/更新时传入，返回时隐藏）
        model_name:
          type: string
          description: 实际模型名
        is_default:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ---------- 提供商相关 ----------
    ProviderType:
      type: string
      enum:
        - openai
        - volcengine
        - deepseek
        - anthropic
        - google
        - moonshot
        - zhipu
        - custom
      description: |
        提供商类型：
        - openai: OpenAI GPT 系列
        - volcengine: 火山引擎 Doubao
        - deepseek: 深度求索
        - anthropic: Anthropic Claude
        - google: Google Gemini
        - moonshot: 月之暗面 Kimi
        - zhipu: 智谱 GLM
        - custom: 自定义 OpenAI 兼容接口
    Provider:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: 用户自定义显示名称（如 "OpenAI (个人)"）
        type:
          $ref: '#/components/schemas/ProviderType'
        api_base:
          type: string
          description: API 基础地址
        models:
          type: array
          items:
            $ref: '#/components/schemas/ProviderModel'
          description: 关联的模型列表
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProviderModel:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider_id:
          type: string
          format: uuid
        model_id:
          type: string
          description: API 返回的模型标识（如 gpt-4o）
        display_name:
          type: string
          description: 用户友好的显示名称
        is_default:
          type: boolean
          description: 是否为该提供商的默认模型
        enabled:
          type: boolean
          description: 是否启用
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AvailableModel:
      type: object
      properties:
        id:
          type: string
          description: 模型标识（如 gpt-4o）
        name:
          type: string
          description: 模型名称（如 GPT-4o）
        owned_by:
          type: string
          description: 模型所有者
...
paths:
...
  /providers:
    get:
      tags:
        - Providers
      summary: 获取提供商列表
      operationId: listProviders
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Provider'

    post:
      tags:
        - Providers
      summary: 创建提供商
      operationId: createProvider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - type
                - api_base
                - api_key
              properties:
                name:
                  type: string
                  description: 显示名称
                  example: OpenAI (个人)
                type:
                  $ref: '#/components/schemas/ProviderType'
                api_base:
                  type: string
                  description: API 地址
                  example: https://api.openai.com/v1
                api_key:
                  type: string
                  description: API Key
                  example: sk-xxx
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    $ref: '#/components/schemas/Provider'
  /providers/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Providers
      summary: 获取提供商详情
      operationId: getProvider
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Provider'

    put:
      tags:
        - Providers
      summary: 更新提供商
      operationId: updateProvider
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: OpenAI (工作)
                api_base:
                  type: string
                  example: https://api.openai.com/v1
                api_key:
                  type: string
                  description: 留空保持不变
                  example: sk-xxx
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  provider:
                    $ref: '#/components/schemas/Provider'

    delete:
      tags:
        - Providers
      summary: 删除提供商
      operationId: deleteProvider
      description: 删除提供商及其所有关联模型
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /providers/{id}/available-models:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Providers
      summary: 获取可用模型列表
      operationId: getAvailableModels
      description: 动态从提供商 API 获取可用模型列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/AvailableModel'
                  is_predefined:
                    type: boolean
                    description: 是否为预定义列表（非动态获取）
        '400':
          description: API 连接失败或认证错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /providers/{providerId}/models:
    parameters:
      - name: providerId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - ProviderModels
      summary: 获取提供商的模型列表
      operationId: listProviderModels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProviderModel'

    post:
      tags:
        - ProviderModels
      summary: 添加模型到提供商
      operationId: addProviderModel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - model_id
              properties:
                model_id:
                  type: string
                  description: 模型标识
                  example: gpt-4o
                display_name:
                  type: string
                  description: 显示名称（可选）
                  example: GPT-4o
                is_default:
                  type: boolean
                  description: 是否设为默认
                  default: false
      responses:
        '201':
          description: 添加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  model:
                    $ref: '#/components/schemas/ProviderModel'
...
  /models:
    deprecated: true
    x-deprecated-message: 请使用 /providers API
    get:
      tags:
        - Models
      summary: 获取模型列表
      operationId: listModels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Model'

    post:
      tags:
        - Models
      summary: 添加模型配置
      operationId: createModel
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - api_base
                - api_key
                - model_name
              properties:
                name:
                  type: string
                  description: 显示名称
                  example: Claude 3.5 Sonnet
                api_base:
                  type: string
                  description: API 地址
                  example: https://api.anthropic.com
                api_key:
                  type: string
                  description: API Key
                  example: sk-ant-xxx
                model_name:
                  type: string
                  description: 实际模型名
                  example: claude-3-5-sonnet-20241022
                is_default:
                  type: boolean
                  description: 是否设为默认
                  default: false
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Model'
...